+++ pipelines/generic.py
@@ -2,6 +2,7 @@ from pipelines.generic_transformer import load_transformer
 from pipelines.generic_text_encoder import load_text_encoder
 from pipelines.generic_vae import load_vae_override
 from pipelines.generic_util import get_loader, set_pipeline
+from pipelines.generic_map import transformers_map
 
 
 __all__ = [
@@ -10,4 +11,5 @@ __all__ = [
     "load_vae_override",
     "get_loader",
     "set_pipeline",
+    "transformers_map",
 ]