+++ pipelines/generic_map.py
@@ -1,0 +1,6 @@
+transformers_map = {
+    'AnimaTextToImagePipeline': 'vladmandic/Anima-1.0-Base',
+    'Krea2Transformer2DModel': 'CalamitousFelicitousness/Krea-2-Base-Diffusers',
+    'CosmosTransformer3DModel': 'vladmandic/Anima-1.0-Base',
+    'ZImageTransformer2DModel': 'Tongyi-MAI/Z-Image',
+}